# navigator-backend/swagger.yml
openapi: 3.0.0
info:
  description: This is the API specification for the Navigator API.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: Navigator
  version: 0.0.1
externalDocs:
  description: Discover the Navigator wallet.
  url: http://navigator.space
servers:
- url: http://api.navigator.space
tags:
- description: Query your Balances
  name: balances
- description: Verify the denom of an IBC token
  name: verify-trace
paths:
  /balances/{addresses*}:
    get:
      description: Returns a list of balances based on the input addresses
      operationId: getBalancesByAddresses
      parameters:
      - description: List of input addresses
        explode: true
        in: path
        name: addresses*
        required: true
        schema:
          type: array
          items:
            type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceArray'
          description: successful operation
        "400":
          description: Invalid address(es) supplied
        "404":
          description: Addresses not found
      security:
      - api_key: []
      summary: Find balances by addresses
      tags:
      - balances
  /verify-trace/{chain}/{denom}:
    get:
      description: Verify the trace of an IBC token on a chain 
      operationId: verifyTrace
      parameters:
      - description: The chain of the token
        explode: false
        in: path
        name: chain
        required: true
        schema:
          type: string
        style: simple
      - description: the IBC denom of a token
        explode: false
        in: path
        name: denom
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TraceResponse'
          description: successful operation
        "400":
          description: Invalid request data
        "404":
          description: Denom not found
      security:
      - api_key: []
      summary: Find balances by addresses
      tags:
      - verify-trace
components:
  schemas:
    Balance:
      type: object
      properties:
        address:
          type: string
        base_denom:
          type: string
        verified:
          type: boolean
        native:
          type: boolean
        amount:
          type: number
        on_chain:
          type: string
        fee_token:
          type: boolean
        ibc:
          type: object
          properties: 
            source_chain:
              type: string
            ibc_denom:
              type: string
            path:
              type: string
            verified_path:
              type: array
              items:
                type: string
      example:
        address: cosmos14pmvh0d4fucylhawvcd0hxkrky99hwcnm0usr5
        base_denom: uluna
        verified: true
        native: false
        amount: 10000
        on_chain: cosmos
        fee_token: true
        ibc:
          source_chain: terra-columbus-3
          ibc_denom: ibc/7F1D3FCF4AE79E1554D670D1AD949A9BA4E4A3C76C63093E17E446A46061A7A2
          path: transfer/channel-0/transfer/channel-6
          verified_path:
          - Kava
          - Terra
    BalanceArray:
      type: array
      items:
        $ref: '#/components/schemas/Balance'
    TraceResponse:
      type: object
      properties:
        ibc_denom:
          type: string
        base_denom:
          type: string
        verified:
          type: boolean
        native:
          type: boolean
        verified_path:
          type: string
        trace:
          type: array
          items:
            $ref: '#/components/schemas/Trace'
    Trace:
      type: object
      properties:
        path:
          type: string
        client_id:
          type: string
        chain_id:
          type: boolean
        chain_name:
          type: boolean